Data Collection [Arduino]
Program collects data from the 64 Hall Effect Sensors installed on each square of the board. MCP23017 was used as an IO expander to allow Arduino to access all 64 pins. Information would be gathered and printed to Serial to allow for the Data Collection Stage to take place (see Data Collection Code [Python]).